Update menu labels for simplicity
The menu labels in the '_data/menu.yml' file have been modified to reduce verbosity and enhance clarity. The 'GitHub Account' label was shortened to 'github', and the 'Email' label, which included the email address, was simplified to 'contact'. The 'Another contents' entry, which was not linked to anything, has been removed as well.