Add post about Chrome extension release
A new post has been added about the release of a Chrome extension that makes it easier to use the Cambridge Dictionary. The post includes an overview of the extension, a link to the source code, a guide on how to use the extension, and a section explaining who would benefit from using the extension.